Add deprecation warnings to "datasets" public interface
