+++ src/phoenix/__init__.py
@@ -4,6 +4,7 @@ from importlib.machinery import ModuleSpec
 from types import ModuleType
 from typing import Any, Optional
 
+from .datasets.dataset import Dataset
 from .inferences.fixtures import ExampleInferences, load_example
 from .inferences.inference import Inference
 from .inferences.schema import EmbeddingColumnNames, RetrievalEmbeddingColumnNames, Schema
@@ -32,6 +33,7 @@ Here are just a few of the things that phoenix does well:
 
 __all__ = [
     "__version__",
+    "Dataset",
     "Inference",
     "EmbeddingColumnNames",
     "RetrievalEmbeddingColumnNames",
