+++ src/phoenix/datasets/fixtures.py
@@ -1,0 +1,21 @@
+from phoenix.inferences.fixtures import (
+    ExampleInferences as _ExampleInferences,
+)
+from phoenix.inferences.fixtures import (
+    load_example as _load_example,
+)
+from phoenix.utilities.deprecation import deprecated, deprecated_class
+
+
+@deprecated_class(
+    "The phoenix.datasets.fixtures module is deprecated, use phoenix.inferences.fixtures instead."
+)
+class ExampleInferences(_ExampleInferences):
+    pass
+
+
+@deprecated(
+    "The phoenix.datasets.fixtures module is deprecated, use phoenix.inferences.fixtures instead."
+)
+def load_example() -> None:
+    _load_example()
