+++ src/phoenix/inferences/inference.py
@@ -1,10 +1,7 @@
 import logging
-import re
 import uuid
 from copy import deepcopy
-from dataclasses import dataclass, fields, replace
-from enum import Enum
-from itertools import groupby
+from dataclasses import fields, replace
 from typing import Any, Dict, List, Optional, Set, Tuple, Union
 
 import numpy as np
@@ -25,7 +22,6 @@ from .schema import (
     SINGLE_COLUMN_SCHEMA_FIELD_NAMES,
     EmbeddingColumnNames,
     EmbeddingFeatures,
-    RetrievalEmbeddingColumnNames,
     Schema,
     SchemaFieldName,
     SchemaFieldValue,
@@ -121,160 +117,6 @@ class Inference:
             schema_json = schema_file.read()
         schema = Schema.from_json(schema_json)
         return cls(df, schema, name)
-
-    @classmethod
-    def from_open_inference(cls, dataframe: DataFrame) -> "Inference":
-        schema = Schema()
-        column_renaming: Dict[str, str] = {}
-        for group_name, group in groupby(
-            sorted(
-                map(_parse_open_inference_column_name, dataframe.columns),
-                key=lambda column: column.name,
-            ),
-            key=lambda column: column.name,
-        ):
-            open_inference_columns = list(group)
-            if group_name == "":
-                column_names_by_category = {
-                    column.category: column.full_name for column in open_inference_columns
-                }
-                schema = replace(
-                    schema,
-                    prediction_id_column_name=column_names_by_category.get(
-                        OpenInferenceCategory.id
-                    ),
-                    timestamp_column_name=column_names_by_category.get(
-                        OpenInferenceCategory.timestamp
-                    ),
-                )
-                continue
-            column_names_by_specifier = {
-                column.specifier: column.full_name for column in open_inference_columns
-            }
-            if group_name == "response":
-                response_vector_column_name = column_names_by_specifier.get(
-                    OpenInferenceSpecifier.embedding
-                )
-                if response_vector_column_name is not None:
-                    column_renaming[response_vector_column_name] = "response"
-                    schema = replace(
-                        schema,
-                        response_column_names=EmbeddingColumnNames(
-                            vector_column_name=column_renaming[response_vector_column_name],
-                            raw_data_column_name=column_names_by_specifier.get(
-                                OpenInferenceSpecifier.default
-                            ),
-                        ),
-                    )
-                else:
-                    response_text_column_name = column_names_by_specifier.get(
-                        OpenInferenceSpecifier.default
-                    )
-                    if response_text_column_name is None:
-                        raise ValueError(
-                            "invalid OpenInference format: missing text column for response"
-                        )
-                    column_renaming[response_text_column_name] = "response"
-                    schema = replace(
-                        schema,
-                        response_column_names=column_renaming[response_text_column_name],
-                    )
-            elif group_name == "prompt":
-                prompt_vector_column_name = column_names_by_specifier.get(
-                    OpenInferenceSpecifier.embedding
-                )
-                if prompt_vector_column_name is None:
-                    raise ValueError(
-                        "invalid OpenInference format: missing embedding vector column for prompt"
-                    )
-                column_renaming[prompt_vector_column_name] = "prompt"
-                schema = replace(
-                    schema,
-                    prompt_column_names=RetrievalEmbeddingColumnNames(
-                        vector_column_name=column_renaming[prompt_vector_column_name],
-                        raw_data_column_name=column_names_by_specifier.get(
-                            OpenInferenceSpecifier.default
-                        ),
-                        context_retrieval_ids_column_name=column_names_by_specifier.get(
-                            OpenInferenceSpecifier.retrieved_document_ids
-                        ),
-                        context_retrieval_scores_column_name=column_names_by_specifier.get(
-                            OpenInferenceSpecifier.retrieved_document_scores
-                        ),
-                    ),
-                )
-            elif OpenInferenceSpecifier.embedding in column_names_by_specifier:
-                vector_column_name = column_names_by_specifier[OpenInferenceSpecifier.embedding]
-                column_renaming[vector_column_name] = group_name
-                embedding_feature_column_names = schema.embedding_feature_column_names or {}
-                embedding_feature_column_names.update(
-                    {
-                        group_name: EmbeddingColumnNames(
-                            vector_column_name=column_renaming[vector_column_name],
-                            raw_data_column_name=column_names_by_specifier.get(
-                                OpenInferenceSpecifier.raw_data
-                            ),
-                            link_to_data_column_name=column_names_by_specifier.get(
-                                OpenInferenceSpecifier.link_to_data
-                            ),
-                        )
-                    }
-                )
-                schema = replace(
-                    schema,
-                    embedding_feature_column_names=embedding_feature_column_names,
-                )
-            elif len(open_inference_columns) == 1:
-                open_inference_column = open_inference_columns[0]
-                raw_column_name = open_inference_column.full_name
-                column_renaming[raw_column_name] = open_inference_column.name
-                if open_inference_column.category is OpenInferenceCategory.feature:
-                    schema = replace(
-                        schema,
-                        feature_column_names=(
-                            (schema.feature_column_names or []) + [column_renaming[raw_column_name]]
-                        ),
-                    )
-                elif open_inference_column.category is OpenInferenceCategory.tag:
-                    schema = replace(
-                        schema,
-                        tag_column_names=(
-                            (schema.tag_column_names or []) + [column_renaming[raw_column_name]]
-                        ),
-                    )
-                elif open_inference_column.category is OpenInferenceCategory.prediction:
-                    if open_inference_column.specifier is OpenInferenceSpecifier.score:
-                        schema = replace(
-                            schema,
-                            prediction_score_column_name=column_renaming[raw_column_name],
-                        )
-                    if open_inference_column.specifier is OpenInferenceSpecifier.label:
-                        schema = replace(
-                            schema,
-                            prediction_label_column_name=column_renaming[raw_column_name],
-                        )
-                elif open_inference_column.category is OpenInferenceCategory.actual:
-                    if open_inference_column.specifier is OpenInferenceSpecifier.score:
-                        schema = replace(
-                            schema,
-                            actual_score_column_name=column_renaming[raw_column_name],
-                        )
-                    if open_inference_column.specifier is OpenInferenceSpecifier.label:
-                        schema = replace(
-                            schema,
-                            actual_label_column_name=column_renaming[raw_column_name],
-                        )
-            else:
-                raise ValueError(f"invalid OpenInference format: duplicated name `{group_name}`")
-
-        return cls(
-            dataframe.rename(
-                column_renaming,
-                axis=1,
-                copy=False,
-            ),
-            schema,
-        )
 
     def to_disc(self) -> None:
         """writes the data and schema to disc"""
@@ -675,51 +517,6 @@ def _get_schema_from_unknown_schema_param(schemaLike: SchemaLike) -> Schema:
 
 def _add_prediction_id(num_rows: int) -> List[str]:
     return [str(uuid.uuid4()) for _ in range(num_rows)]
-
-
-class OpenInferenceCategory(Enum):
-    id = "id"
-    timestamp = "timestamp"
-    feature = "feature"
-    tag = "tag"
-    prediction = "prediction"
-    actual = "actual"
-
-
-class OpenInferenceSpecifier(Enum):
-    default = ""
-    score = "score"
-    label = "label"
-    embedding = "embedding"
-    raw_data = "raw_data"
-    link_to_data = "link_to_data"
-    retrieved_document_ids = "retrieved_document_ids"
-    retrieved_document_scores = "retrieved_document_scores"
-
-
-@dataclass(frozen=True)
-class _OpenInferenceColumnName:
-    full_name: str
-    category: OpenInferenceCategory
-    data_type: str
-    specifier: OpenInferenceSpecifier = OpenInferenceSpecifier.default
-    name: str = ""
-
-
-def _parse_open_inference_column_name(column_name: str) -> _OpenInferenceColumnName:
-    pattern = (
-        r"^:(?P<category>\w+)\.(?P<data_type>\[\w+\]|\w+)(\.(?P<specifier>\w+))?:(?P<name>.*)?$"
-    )
-    if match := re.match(pattern, column_name):
-        extract = match.groupdict(default="")
-        return _OpenInferenceColumnName(
-            full_name=column_name,
-            category=OpenInferenceCategory(extract.get("category", "").lower()),
-            data_type=extract.get("data_type", "").lower(),
-            specifier=OpenInferenceSpecifier(extract.get("specifier", "").lower()),
-            name=extract.get("name", ""),
-        )
-    raise ValueError(f"Invalid format for column name: {column_name}")
 
 
 # A dataset with no data. Useful for stubs
